analysis/similarity_search/PLM_Embedding/GeneAnnotation/requirements.txt
pymilvus
numpy
transformers
torch
biopython
pandas
sentencepiece
openai
